Correctly hash UDP Payload
The UDP->Length includes the UDP header and payload.  If we want to hash
only the Payload then we must subtract the size of the UDP header.

Additionally this fixes the UDP Length Chomp which should futher reduce
the size of the payload that is hashed, not extend it past the end of
the packet.

Signed-off-by: Shawn Bohrer <sbohrer@rgmadvisors.com>